+++ app/models/project.rb
@@ -42,7 +42,7 @@ class Project < ActiveRecord::Base
 
   has_many :project_favorites
 
-  has_many :adverse_events, -> { current }
+  has_many :adverse_events, -> { current.joins(:subject).merge(Subject.current) }
   has_many :categories, -> { where(deleted: false).order(:position) }
   has_many :designs, -> { where deleted: false }
   has_many :variables, -> { where deleted: false }