+++ db/migrate/20160210202205_add_include_adverse_events_to_exports.rb
@@ -1,0 +1,5 @@
+class AddIncludeAdverseEventsToExports < ActiveRecord::Migration
+  def change
+    add_column :exports, :include_adverse_events, :boolean, null: false, default: false
+  end
+end